veilcore/veil_os/pyproject.toml
[project]
name = "veil-os"
version = "0.1.0"
description = "Veil OS hybrid organ engine"
authors = [{ name = "Future_Ready" }]
requires-python = ">=3.10"
dependencies = ["pyyaml", "fastapi", "uvicorn", "jinja2"]

[project.scripts]
veil = "veil.cli:main"
veil-compile = "veil.compiler:compile_all"
veil-hospital = "veil.hospital.__main__:main"
veil-intrusion = "veil.intrusion.__main__:main"
veil-api = "veil.api.__main__:main"
veil-sentinel = "veil.organs.sentinel.runner:main"
veil-insider-threat = "veil.organs.insider_threat.runner:main"
veil-auto-lockdown = "veil.organs.auto_lockdown.runner:main"
veil-zero-trust = "veil.organs.zero_trust.runner:main"

[tool.setuptools]
packages = ["veil", "veil.organs", "veil.organs.sentinel", "veil.organs.insider_threat", "veil.organs.auto_lockdown", "veil.organs.zero_trust", "veil.security"]
